CSP 5.8.0 (#590)
* [NT-0] doc: Assets (#576)

This change introduces a new module to the CSP documentation site: 'Assets'.

The module covers the following topics:

Understanding Asset Collections
Asset CRUD Operations
Relationship with Entity Components
Querying for Assets
Automated Model Decimation

* [OF-1437] refac: Introduce a new Event Bus for SignalR Multiplayer events (#533)

- create EventBus class and move event handling to it
- separate NetworkEventsMap into two, one for Systems
  and one for Callbacks
- update the SystemBase base class to handle default
  callbacks and serialization
- make default SystemBase constructor private as it only
  exists to appease the wrapper generator
- move System callbacks and deserializers (OnEvent) to their
  respective System: Systems can register to events by
  passing their instance; clients can register with a
  parameterised callback as before
- update tests to work with the new EventBus
- move event tests from MultiPlayerTests.cpp
  to new EventBusTests.cpp and add test case for the new 
 Callback vs Systems functionality
- add checks for expected log messages to new test
- fix wrapper generator enum parsing

* [OF-1471] test: Integrate multiplayer test runner with Tests project (#575)

* [OF-1471] test: Add tiny-process-lib

Adds tiny-process-lib, for the purposes
of the multiplayer test runner, and
invoking it within the tests.

I laboured over making this a submodule vs
having it directly in thirdparty, and eventually
came to the conclusion that direct thirdparty
is more standard to the codebase.

* [OF-1471] test: Build & Install TPL

Build TPL using cmake during the solution creation.
Copy it to the directory of the tests so its easily
invokable.

We go to the effort to install tiny-process-lib to
an install directory here, which isn't something we
do for other cmake libs, but really we should.

* [OF-1471] test: Dont test MTR by default

During dev it seemed cute to have debug builds
run the tests by default, but it immediately became
frustrating during integration.

* [OF-1471] test: Test Runner Shakedowns

A couple of changes to the test runner that emerged
during integration.

1. I hit the classic problem of calling a private header
in a public one.

2. The process descriptors were too generic,
and I actually accidentlyoutput the same text during
debugging, which caused pain due to
how we monitor stdout to flag futures.

* [OF-1471] test: MTR process manager

Add the process controller object for use in our tests.

Allows setting command line arguments and invoking
the multiplayer test runner, and managing termination
via RAII.

The manager works via listening for set strings on
stdout, and then setting promises to notify the main
test of when events have occured on each process.

* [OF-1471] test: MTR integration test

Add an integration test that launches the
multiplayer test runner and checks that
its operations are performed.

* [OF-1471] test: Unit tests for runner process

Unit tests for the runner process object itself.
Not tests for the runner itself, which has its own tests,
(although they're not run automatically yet / ever)

* [NT-0] doc: updated stack graphic (#581)

Updating the apps/CSP/services stack graphic with a new visual treatment.

* [NT-0] doc: visual treatment for diagrams (#586)

Updated various diagrams throughout the documentation with a new visual treatment.

* [NT-0] style: In hooks, Remove line limitations from both git commit title and git commit body

I am intentionally writing an overly long subject line and overly long message to test that this works.

* [OF-1527] Allow test endpoint to be configurable (#585)

Exposed the services endpoint to be configurable or default to live services if not defined.

* [OF-1495] style: update clang-format (#587)

- Update clang-format to use the WebKit preset style
- Set the number of columns to 150 (as it was before)
- Use the Allman brace breaking style (as it was before)
- Run new clang-format on all *.cpp/*.hpp/*.c/*.h files,
  excluding modules and third parties

* [OF-1492] test: Review all tests and either put them on the CI or delete them (#583)

- Re-enable commented out/disabled C++ tests that pass, and use
  DISABLE_CSPEngine for the tests that fail/that take 40 minutes
- Delete C++ tests that are no longer needed
- Update C++ tests that no longer build
- Add link to story for every C++ test that has one
- Remove all CSharp tests, as these are no longer run nor 
  maintained, except the Wrapper Generator UnitTesting ones, which
  we run on the CI
- Remove all JavaScript tests, as these are no longer run nor
  maintained, except the Wrapper Generator UnitTesting ones, which
  we run as a GitHub Action

* [OF-1441] feat: Video Material Classes (#530)

* [OF-1441] feat: added material classes

* [OF-1441] feat: updated defult values

* [OF-1441] fix: fixed enums for wrapper gen

* [OF-1441] feat: added IsSet for textures

* [OF-1441] style: updated comments

* [OF-1441] feat: adding material base class

* [OF-1441] fix: Set textures when deserializing

* [OF-1440] feat: Material AssetSystem Interface (#547)

* [OF-1441] fix: Set textures when deserializing

* [OF-1440] feat: add asset system interface

* [OF-1440] feat: Add Asset interface for materials

* [OF-1440] test: Add Material tests

* [OF-1440] refac: inline docs + interface update

* [OF-1440] fix: make assetcollectionid unique

* [OF-1440] feat: creating unique asset names

* [OF-1440] fix: Fix GetMaterials + style

* [OF-1440] refac: pr reviews

* [OF-1440] refac: remove const cast

* [OF-1442] feat: Material Overrides (#540)

* [OF-1441] fix: Set textures when deserializing

* [OF-1442] feat: add material overrides

* Update Library/include/CSP/Multiplayer/Components/StaticModelSpaceComponent.h


* [OF-1442] refac: pr reviews

---------


* [OF-1441] fix: name changes

* [OF-1441] refac: update to new csp api

* [OF-1408] fix: update AvatarInfoResult type (#566)

* [NT-0] doc: Physical and Digital Realities (#565)

This change introduces a new module to the CSP documentation site: 'Physical and Digital Realities'.

The module covers the following topics:

Cross-Reality Applications (and digital twins)
Anchoring (with a focus on how to use CSP and Google Cloud Anchors)
Fiducial Markers

* [OB-3643] fix: update callbacks

* Ensured callbacks are now called on main thread

* Revert "[OB-3643] fix: update callbacks"

* [OF-1519] fix: replicated map (#573)

* [OF-1519] fix: update to string map

* [OF-1519] fix: update map serialization

* Map serialization has been changed to use a string map

* [OF-1519] test: add tests

* [OF-1519] refac: remove test vars

* [OF-1519] refac: fix typo

* [OF-1519] test: update tests

* [OF-1441] fix: main thread crash

* [OF-1441] fix: remove pthread proxy

* [OF-1441] fix: wasm crash (#578)

* [OF-1441] fix: wasm crash

* [OF-1441] refac: simplify material downloading

* [OF-1518] feat: add material change event (#580)

* [OF-1441] doc: update comment

* [OF-1441] fix: elegantly fail when error (#582)

* [OF-1441] fix: elegantly fail when error

* [OF-1441] doc: Add comment

* [OF-1441] fix: update to new login method

* [OF-1441] test: update to new login

---------


---------

+++ Library/src/Systems/ECommerce/ECommerce.cpp
@@ -673,8 +673,8 @@ void ProductInfoCollectionResult::OnResponse(const csp::services::ApiResponseBas
         std::vector<chs_aggregation::ShopifyProductDto>& ProductsArray = ProductCollectionResponse->GetArray();
 
         // Loop through products to count the variants, we want 1 output product per variant
-        size_t VariantCount = 0;
-        for (size_t DtoCount = 0; DtoCount < ProductsArray.size(); DtoCount++)
+        int VariantCount = 0;
+        for (int DtoCount = 0; DtoCount < ProductsArray.size(); DtoCount++)
         {
             VariantCount += ProductsArray[DtoCount].GetVariants().size();
         }
